requirements.txt
fastapi
pydantic
uvicorn
fastapi-mail
python-dotenv